+++ HISTORY.md
@@ -37,6 +37,7 @@
   @mmoisse for the bug report.
 - Add support for peddy (https://github.com/brentp/peddy) for PED file 
   correspondence/ancestry checking.
+- seq2c: pass through mapping_reads.txt file to upload directory.
 
 ## 1.0.6 (5 November 2017)